+++ drools-planner-core/src/main/java/org/drools/planner/core/heuristic/selector/move/composite/UnionMoveSelector.java
@@ -122,7 +122,7 @@ public class UnionMoveSelector extends CompositeMoveSelector {
     public long getSize() {
         long size = 0L;
         for (MoveSelector moveSelector : childMoveSelectorList) {
-            size += (long) moveSelector.getSize();
+            size += moveSelector.getSize();
         }
         return size;
     }
